% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/load_la.R
\name{load_la}
\alias{load_la}
\title{London's local authority boundaries}
\usage{
load_la()
}
\value{
Object 'lon', a 33 row SpatialPolygonsDataFrame
}
\description{
Load shapefile of London's local authority boundaries (GLA boroughs) as a SpatialPolygonsDataFrame
}
\examples{
load_la.R()
}

