% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/load_lon.R
\name{load_lon}
\alias{load_lon}
\title{London's outline}
\usage{
load_lon()
}
\value{
Object 'lon', a single row SpatialPolygonsDataFrame
}
\description{
Load shapefile of London's outline (Greater London Authority) as a SpatialPolygonsDataFrame
}
\examples{
load_lon.R()
}

