% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/load_thames.R
\name{load_thames}
\alias{load_thames}
\title{River Thames}
\usage{
load_thames()
}
\value{
Object 'lon', a 1 row SpatialPolygonsDataFrame
}
\description{
Load shapefile of River Thames as a SpatialPolygonsDataFrame
}
\examples{
load_thames.R()
}

