\name{londonShapefiles-package}
\alias{londonShapefiles-package}
\alias{londonShapefiles}
\docType{package}
\title{
Some shapefiles for London
}
\description{
Some useful (small) boundary shapefiles, including GLA boundary, local authorities and River Thames
}
\details{
\tabular{ll}{
Package: \tab londonShapefiles\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2015-03-10\cr
License: \tab GPL-2\cr
}
Call functions to load the individual shapefiles
}
\author{
Robin Edwards

Maintainer: Robin Edwards <geotheory1@gmail.com>
}
\references{
}
\keyword{ package }
\examples{
load_lon()     # London outline
load_la()      # local authority boundaries
load_thames()  # River Thames
}
